Add files via upload
This is the second version of my upload and not too much has changed. I have just modified it slightly to make it more efficient in terms of the amount of space the code takes up and refined it, in order to increase readability when actually running the programme. For example, commas are now inserted when numbers become greater than or equal to x 10^4 e.g originally 2345659 will read 2,345,659.