Add files via upload
This version of the programme now does quite a lot more. It can compute fibonacci numbers, most arithmetic and quadratic sequences and finds the nth term for them. Furthermore, better axis and titles of cumulative sequence graphs have been made with the nth term showcased. The program will now print a message at the bottom when terms are printed, stating how many terms are there from 0 to your upper boundary for the kind of sequence it is e.g. the prime numbers or 3, 8, 15, 24, 35... When using the programme make sure to download both files as the code will not work otherwise; to run it, do so from the 'Plotting_Printing.py' file. Finally, when inputting a non-special sequence please include at least 5 terms of the sequence each separated by a comma followed by a space otherwise it will not work. Merry Christmas.